requirements.txt
aioconsole==0.4.1
aiohttp==3.8.1
emoji==1.7.0
nextcord==2.0.0b4
pre-commit==2.19.0
PyNaCl==1.5.0
python-dotenv==0.20.0
yt-dlp==2022.5.18
